ADD SettingBooks.js
